Move functions in views.py that do not return an HTTP response to name/utils.py.
